+++ lib/auto_increment/active_record.rb
@@ -4,7 +4,7 @@ module AutoIncrement
   module ActiveRecord
 
     def auto_increment(column = nil, options = {})
-      options.reverse_merge!(before: :create)
+      options = options.reverse_merge(before: :create)
       callback = "before_#{options[:before]}"
       send callback, Incrementor.new(column, options)
     end